+++ src/argilla/server/apis/v1/handlers/datasets.py
@@ -13,25 +13,103 @@
 #  limitations under the License.
 
 from typing import List
+from uuid import UUID
 
-from fastapi import APIRouter, Depends, Security
+from fastapi import APIRouter, Depends, HTTPException, Security, status
 from sqlalchemy.orm import Session
 
 from argilla.server.contexts import datasets
 from argilla.server.database import get_db
 from argilla.server.policies import DatasetPolicyV1, authorize
-from argilla.server.schemas.v1.datasets import Dataset
+from argilla.server.schemas.v1.datasets import Annotation, Dataset, DatasetCreate
 from argilla.server.security import auth
 from argilla.server.security.model import User
 
 router = APIRouter(tags=["datasets"])
 
 
+def _get_dataset(db: Session, dataset_id: UUID):
+    dataset = datasets.get_dataset_by_id(db, dataset_id)
+    if not dataset:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Dataset with id `{dataset_id}` not found",
+        )
+
+    return dataset
+
+
 @router.get("/datasets", response_model=List[Dataset])
-def list_datasets(*, db: Session = Depends(get_db), current_user: User = Security(auth.get_current_user)):
+def list_datasets(
+    *,
+    db: Session = Depends(get_db),
+    current_user: User = Security(auth.get_current_user),
+):
     authorize(current_user, DatasetPolicyV1.list)
 
     if current_user.is_admin:
         return datasets.list_datasets(db)
     else:
         return current_user.datasets
+
+
+@router.get("/datasets/{dataset_id}", response_model=Dataset)
+def get_dataset(
+    *,
+    db: Session = Depends(get_db),
+    dataset_id: UUID,
+    current_user: User = Security(auth.get_current_user),
+):
+    dataset = _get_dataset(db, dataset_id)
+
+    authorize(current_user, DatasetPolicyV1.get(dataset))
+
+    return dataset
+
+
+@router.get("/datasets/{dataset_id}/annotations", response_model=List[Annotation])
+def get_dataset_annotations(
+    *,
+    db: Session = Depends(get_db),
+    dataset_id: UUID,
+    current_user: User = Security(auth.get_current_user),
+):
+    dataset = _get_dataset(db, dataset_id)
+
+    authorize(current_user, DatasetPolicyV1.get(dataset))
+
+    return dataset.annotations
+
+
+@router.post("/datasets", status_code=status.HTTP_201_CREATED, response_model=Dataset)
+def create_dataset(
+    *,
+    db: Session = Depends(get_db),
+    dataset_create: DatasetCreate,
+    current_user: User = Security(auth.get_current_user),
+):
+    authorize(current_user, DatasetPolicyV1.create)
+
+    if datasets.get_dataset_by_name_and_workspace_id(db, dataset_create.name, dataset_create.workspace_id):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"Dataset with name `{dataset_create.name}` already exists for workspace with id `{dataset_create.workspace_id}`",
+        )
+
+    return datasets.create_dataset(db, dataset_create)
+
+
+@router.delete("/datasets/{dataset_id}", response_model=Dataset)
+def delete_dataset(
+    *,
+    db: Session = Depends(get_db),
+    dataset_id: UUID,
+    current_user: User = Security(auth.get_current_user),
+):
+    authorize(current_user, DatasetPolicyV1.delete)
+
+    dataset = _get_dataset(db, dataset_id)
+
+    datasets.delete_dataset(db, dataset)
+
+    return dataset