+++ src/argilla/server/schemas/v1/datasets.py
@@ -18,12 +18,34 @@ from uuid import UUID
 
 from pydantic import BaseModel
 
+from argilla.server.models import AnnotationType
+
 
 class Dataset(BaseModel):
     id: UUID
     name: str
     guidelines: Optional[str]
     workspace_id: UUID
+    inserted_at: datetime
+    updated_at: datetime
+
+    class Config:
+        orm_mode = True
+
+
+class DatasetCreate(BaseModel):
+    name: str
+    guidelines: Optional[str]
+    workspace_id: UUID
+
+
+class Annotation(BaseModel):
+    id: UUID
+    name: str
+    title: str
+    type: AnnotationType
+    required: bool
+    settings: dict
     inserted_at: datetime
     updated_at: datetime