+++ tests/server/api/v1/test_datasets.py
@@ -12,15 +12,20 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+from datetime import datetime
+from uuid import UUID, uuid4
+
 from argilla._constants import API_KEY_HEADER_NAME
+from argilla.server.models import AnnotationType, Dataset
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
 from tests.factories import (
     AnnotatorFactory,
     DatasetFactory,
+    RatingAnnotationFactory,
+    TextAnnotationFactory,
     WorkspaceFactory,
-    WorkspaceUserFactory,
 )
 
 
@@ -67,9 +72,8 @@ def test_list_datasets_without_authentication(client: TestClient):
 
 
 def test_list_datasets_as_annotator(client: TestClient, db: Session):
-    annotator = AnnotatorFactory.create()
     workspace = WorkspaceFactory.create()
-    WorkspaceUserFactory.create(workspace_id=workspace.id, user_id=annotator.id)
+    annotator = AnnotatorFactory.create(workspaces=[workspace])
 
     DatasetFactory.create(name="dataset-a", workspace=workspace)
     DatasetFactory.create(name="dataset-b", workspace=workspace)
@@ -79,3 +83,222 @@ def test_list_datasets_as_annotator(client: TestClient, db: Session):
 
     assert response.status_code == 200
     assert [dataset["name"] for dataset in response.json()] == ["dataset-a", "dataset-b"]
+
+
+def test_get_dataset(client: TestClient, admin_auth_header: dict):
+    dataset = DatasetFactory.create(name="dataset")
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}", headers=admin_auth_header)
+
+    assert response.status_code == 200
+    assert response.json() == {
+        "id": str(dataset.id),
+        "name": "dataset",
+        "guidelines": None,
+        "workspace_id": str(dataset.workspace_id),
+        "inserted_at": dataset.inserted_at.isoformat(),
+        "updated_at": dataset.updated_at.isoformat(),
+    }
+
+
+def test_get_dataset_without_authentication(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}")
+
+    assert response.status_code == 401
+
+
+def test_get_dataset_as_annotator(client: TestClient, db: Session):
+    dataset = DatasetFactory.create(name="dataset")
+    annotator = AnnotatorFactory.create(workspaces=[dataset.workspace])
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}", headers={API_KEY_HEADER_NAME: annotator.api_key})
+
+    assert response.status_code == 200
+    assert response.json()["name"] == "dataset"
+
+
+def test_get_dataset_as_annotator_from_different_workspace(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+    annotator = AnnotatorFactory.create(workspaces=[WorkspaceFactory.build()])
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}", headers={API_KEY_HEADER_NAME: annotator.api_key})
+
+    assert response.status_code == 403
+
+
+def test_get_dataset_with_nonexistent_dataset_id(client: TestClient, db: Session, admin_auth_header: dict):
+    DatasetFactory.create()
+
+    response = client.get(f"/api/v1/datasets/{uuid4()}", headers=admin_auth_header)
+
+    assert response.status_code == 404
+
+
+def test_get_dataset_annotations(client: TestClient, db: Session, admin_auth_header: dict):
+    dataset = DatasetFactory.create()
+    text_annotation = TextAnnotationFactory.create(
+        name="text-annotation", title="Text Annotation", required=True, dataset=dataset
+    )
+    rating_annotation = RatingAnnotationFactory.create(
+        name="rating-annotation", title="Rating Annotation", dataset=dataset
+    )
+    TextAnnotationFactory.create()
+    RatingAnnotationFactory.create()
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}/annotations", headers=admin_auth_header)
+
+    assert response.status_code == 200
+    assert response.json() == [
+        {
+            "id": str(text_annotation.id),
+            "name": "text-annotation",
+            "title": "Text Annotation",
+            "type": AnnotationType.text.value,
+            "required": True,
+            "settings": {},
+            "inserted_at": text_annotation.inserted_at.isoformat(),
+            "updated_at": text_annotation.updated_at.isoformat(),
+        },
+        {
+            "id": str(rating_annotation.id),
+            "name": "rating-annotation",
+            "title": "Rating Annotation",
+            "type": AnnotationType.rating.value,
+            "required": False,
+            "settings": {},
+            "inserted_at": rating_annotation.inserted_at.isoformat(),
+            "updated_at": rating_annotation.updated_at.isoformat(),
+        },
+    ]
+
+
+def test_get_dataset_annotations_without_authentication(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+
+    response = client.get(f"/api/v1/datasets/{dataset.id}/annotations")
+
+    assert response.status_code == 401
+
+
+def test_get_dataset_annotations_as_annotator(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+    annotator = AnnotatorFactory.create(workspaces=[dataset.workspace])
+    TextAnnotationFactory.create(name="text-annotation", dataset=dataset)
+    RatingAnnotationFactory.create(name="rating-annotation", dataset=dataset)
+    TextAnnotationFactory.create()
+    RatingAnnotationFactory.create()
+
+    response = client.get(
+        f"/api/v1/datasets/{dataset.id}/annotations", headers={API_KEY_HEADER_NAME: annotator.api_key}
+    )
+
+    assert response.status_code == 200
+    assert [annotation["name"] for annotation in response.json()] == ["text-annotation", "rating-annotation"]
+
+
+def test_get_dataset_annotations_as_annotator_from_different_workspace(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+    annotator = AnnotatorFactory.create(workspaces=[WorkspaceFactory.build()])
+
+    response = client.get(
+        f"/api/v1/datasets/{dataset.id}/annotations", headers={API_KEY_HEADER_NAME: annotator.api_key}
+    )
+
+    assert response.status_code == 403
+
+
+def test_get_dataset_annotations_with_nonexistent_dataset_id(client: TestClient, db: Session, admin_auth_header: dict):
+    DatasetFactory.create()
+
+    response = client.get(f"/api/v1/datasets/{uuid4()}/annotations", headers=admin_auth_header)
+
+    assert response.status_code == 404
+
+
+def test_create_dataset(client: TestClient, db: Session, admin_auth_header: dict):
+    workspace = WorkspaceFactory.create()
+    dataset_json = {"name": "name", "guidelines": "guidelines", "workspace_id": str(workspace.id)}
+
+    response = client.post("/api/v1/datasets", headers=admin_auth_header, json=dataset_json)
+
+    assert response.status_code == 201
+    assert db.query(Dataset).count() == 1
+
+    response_body = response.json()
+    assert db.get(Dataset, UUID(response_body["id"]))
+    assert response_body == {
+        "id": str(UUID(response_body["id"])),
+        "name": "name",
+        "guidelines": "guidelines",
+        "workspace_id": str(workspace.id),
+        "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
+        "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
+    }
+
+
+def test_create_dataset_without_authentication(client: TestClient, db: Session):
+    dataset_json = {"name": "name", "workspace_id": str(WorkspaceFactory.create().id)}
+
+    response = client.post("/api/v1/datasets", json=dataset_json)
+
+    assert response.status_code == 401
+    assert db.query(Dataset).count() == 0
+
+
+def test_create_dataset_as_annotator(client: TestClient, db: Session):
+    annotator = AnnotatorFactory.create()
+    dataset_json = {"name": "name", "workspace_id": str(WorkspaceFactory.create().id)}
+
+    response = client.post("/api/v1/datasets", headers={API_KEY_HEADER_NAME: annotator.api_key}, json=dataset_json)
+
+    assert response.status_code == 403
+    assert db.query(Dataset).count() == 0
+
+
+def test_create_dataset_with_existent_name(client: TestClient, db: Session, admin_auth_header: dict):
+    dataset = DatasetFactory.create(name="name")
+    dataset_json = {"name": "name", "workspace_id": str(dataset.workspace_id)}
+
+    response = client.post("/api/v1/datasets", headers=admin_auth_header, json=dataset_json)
+
+    assert response.status_code == 409
+    assert db.query(Dataset).count() == 1
+
+
+def test_delete_dataset(client: TestClient, db: Session, admin_auth_header: dict):
+    dataset = DatasetFactory.create()
+
+    response = client.delete(f"/api/v1/datasets/{dataset.id}", headers=admin_auth_header)
+
+    assert response.status_code == 200
+    assert db.query(Dataset).count() == 0
+
+
+def test_delete_dataset_without_authentication(client: TestClient, db: Session):
+    dataset = DatasetFactory.create()
+
+    response = client.delete(f"/api/v1/datasets/{dataset.id}")
+
+    assert response.status_code == 401
+    assert db.query(Dataset).count() == 1
+
+
+def test_delete_dataset_as_annotator(client: TestClient, db: Session):
+    annotator = AnnotatorFactory.create()
+    dataset = DatasetFactory.create()
+
+    response = client.delete(f"/api/v1/datasets/{dataset.id}", headers={API_KEY_HEADER_NAME: annotator.api_key})
+
+    assert response.status_code == 403
+    assert db.query(Dataset).count() == 1
+
+
+def test_delete_dataset_with_nonexistent_dataset_id(client: TestClient, db: Session, admin_auth_header: dict):
+    DatasetFactory.create()
+
+    response = client.delete(f"/api/v1/datasets/{uuid4()}", headers=admin_auth_header)
+
+    assert response.status_code == 404
+    assert db.query(Dataset).count() == 1